pyproject.toml
[tool.poetry]
name = "nanobanana"
version = "0.1.0"
description = "High-quality Gemini image generation library with adaptive rate limiting"
authors = ["Rob Gruhl"]
readme = "README.md"
packages = [{include = "nanobanana", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
google-genai = "^1.0"
click = "^8.0"
rich = "^13.0"
pydantic = "^2.0"
python-dotenv = "^1.0"
pillow = "^10.0"
pyyaml = "^6.0"

[tool.poetry.scripts]
gemini-image = "nanobanana.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
